Change stake cooldown from seconds to blocks

+++ contracts/protocol/stake/interfaces/IStakeToken.sol
@@ -14,7 +14,7 @@ interface IStakeToken is IDerivedToken {
 
   function cooldown() external;
 
-  function getCooldown(address) external returns (uint40);
+  function getCooldown(address) external returns (uint32);
 
   function exchangeRate() external view returns (uint256);
